+++ packages/fluentui/docs/src/vr-tests/Popup/PopperExamplePositioning.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { StoryWright, Steps } from 'storywright';
-import { ComponentMeta } from '@storybook/react';
+import { Meta } from '@storybook/react';
 import { Popup } from '@fluentui/react-northstar';
 import PopperExamplePositioning from '../../examples/components/Popup/Visual/PopperExamplePositioning';
 
@@ -26,6 +26,6 @@ export default {
       </StoryWright>
     ),
   ],
-} as ComponentMeta<typeof Popup>;
+} as Meta<typeof Popup>;
 
 export { PopperExamplePositioning };