+++ packages/react-components/react-popover/library/src/components/PopoverSurface/usePopoverSurfaceStyles.styles.ts
@@ -22,11 +22,15 @@ const useStyles = makeStyles({
   root: {
     color: tokens.colorNeutralForeground1,
     backgroundColor: tokens.colorNeutralBackground1,
-    boxShadow: tokens.shadow16,
     borderRadius: tokens.borderRadiusMedium,
     border: `1px solid ${tokens.colorTransparentStroke}`,
     ...typographyStyles.body1,
     ...createSlideStyles(10),
+
+    // TODO need to add versions of tokens.alias.shadow.shadow16, etc. that work with filter
+    filter:
+      `drop-shadow(0 0 2px ${tokens.colorNeutralShadowAmbient}) ` +
+      `drop-shadow(0 8px 16px ${tokens.colorNeutralShadowKey})`,
   },
 
   inline: {