+++ packages/react-components/react-avatar/library/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-avatar
 
-This log was last generated on Tue, 30 Jul 2024 18:47:35 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Aug 2024 22:33:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.6.35](https://github.com/microsoft/fluentui/tree/@fluentui/react-avatar_v9.6.35)
+
+Mon, 05 Aug 2024 22:33:04 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-avatar_v9.6.34..@fluentui/react-avatar_v9.6.35)
+
+### Patches
+
+- Bump @fluentui/react-popover to v9.9.17 ([PR #32077](https://github.com/microsoft/fluentui/pull/32077) by beachball)
 
 ## [9.6.34](https://github.com/microsoft/fluentui/tree/@fluentui/react-avatar_v9.6.34)